provisioning SingleActivity principle drawer now inside fragment andr now it is controlling fragments transactions
