+++ app/src/main/AndroidManifest.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Created by Andrii Kovalchuk on 29.11.19 21:45
+  ~ Created by Andrii Kovalchuk on 02.12.19 20:57
   ~ Copyright (c) 2019. All rights reserved.
-  ~ Last modified 29.11.19 17:41
+  ~ Last modified 02.12.19 20:08
   ~
   ~ This Source Code Form is subject to the terms of the Mozilla Public
   ~ License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -25,7 +25,7 @@
 		android:label="@string/app_name"
 		android:roundIcon="@mipmap/ic_launcher_round"
 		android:supportsRtl="true"
-		android:theme="@style/AppTheme.NoActionBar">
+		android:theme="@style/AppTheme">
 
 		<activity
 			android:name=".ui.SplashActivity"
@@ -39,7 +39,7 @@
 		</activity>
 
 		<activity
-			android:name=".ui.main.view.MainActivity"
+			android:name=".ui.MainActivity"
 			android:label="@string/app_name"
 			android:screenOrientation="portrait"
 			android:windowSoftInputMode="adjustPan"/>
