+++ app/src/main/java/com/mmdev/roove/ui/core/BaseFragment.kt
@@ -1,24 +1,25 @@
 /*
- * Created by Andrii Kovalchuk on 20.08.19 13:37
+ * Created by Andrii Kovalchuk on 02.12.19 20:57
  * Copyright (c) 2019. All rights reserved.
- * Last modified 18.11.19 20:01
+ * Last modified 02.12.19 20:53
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package com.mmdev.roove.ui.main.view
+package com.mmdev.roove.ui.core
+
+import androidx.fragment.app.Fragment
 
 /**
  * This is the documentation block about the class
  */
 
-interface MainActivityListeners {
+abstract class BaseFragment (layoutRes: Int = 0) : Fragment(layoutRes) {
 
-	fun startAuthActivity()
-	fun onCardsClick()
 
-	fun onLogOutClick()
+	open fun onBackPressed() {}
+
 
 }
