+++ app/src/main/res/layout/activity_auth.xml
@@ -1,10 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ Created by Andrii Kovalchuk on 02.12.19 20:57
+  ~ Copyright (c) 2019. All rights reserved.
+  ~ Last modified 02.12.19 19:09
+  ~
+  ~ This Source Code Form is subject to the terms of the Mozilla Public
+  ~ License, v. 2.0. If a copy of the MPL was not distributed with this
+  ~ file, You can obtain one at https://mozilla.org/MPL/2.0/.
+  -->
+
 <androidx.constraintlayout.widget.ConstraintLayout
 	android:id="@+id/auth_container"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
 	android:background="@color/colorPrimaryDark"
-	android:paddingTop="24dp"
 	xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:app="http://schemas.android.com/apk/res-auto"
 	xmlns:tools="http://schemas.android.com/tools">
