+++ app/src/main/res/layout/activity_main.xml
@@ -1,75 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Created by Andrii Kovalchuk on 01.12.19 22:42
+  ~ Created by Andrii Kovalchuk on 02.12.19 20:57
   ~ Copyright (c) 2019. All rights reserved.
-  ~ Last modified 01.12.19 22:41
+  ~ Last modified 02.12.19 19:53
   ~
   ~ This Source Code Form is subject to the terms of the Mozilla Public
   ~ License, v. 2.0. If a copy of the MPL was not distributed with this
   ~ file, You can obtain one at https://mozilla.org/MPL/2.0/.
   -->
 
-<androidx.drawerlayout.widget.DrawerLayout
-	android:id="@+id/drawer_layout"
-	android:layout_width="match_parent"
+
+<FrameLayout
+	android:id="@+id/main_activity_container"
 	android:layout_height="match_parent"
-	android:background="@color/colorPrimaryDark"
-	tools:context=".ui.main.view.MainActivity"
-	tools:openDrawer="start"
-	xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:app="http://schemas.android.com/apk/res-auto"
-	xmlns:tools="http://schemas.android.com/tools">
-
-	<androidx.coordinatorlayout.widget.CoordinatorLayout
-		android:id="@+id/main_core_container"
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:background="@color/colorPrimaryDark"
-		android:paddingTop="@dimen/notifbar_size_24">
-
-		<com.google.android.material.appbar.AppBarLayout
-			android:id="@+id/app_bar"
-			android:layout_width="match_parent"
-			android:layout_height="?actionBarSize"
-			android:theme="@style/AppTheme.AppBarOverlay">
-
-			<androidx.appcompat.widget.Toolbar
-				android:id="@+id/toolbar"
-				android:layout_width="match_parent"
-				android:layout_height="match_parent"
-				android:background="@color/colorPrimary"
-				app:layout_scrollFlags="scroll|enterAlways|snap"
-				app:popupTheme="@style/menuPopupStyle"
-				app:titleTextAppearance="@style/regular_text_white_20sp_85op"
-				app:title="@string/app_name"/>
-
-		</com.google.android.material.appbar.AppBarLayout>
-
-		<androidx.constraintlayout.widget.ConstraintLayout
-			android:id="@+id/main_container"
-			android:layout_width="match_parent"
-			android:layout_height="match_parent"
-			android:background="@color/colorPrimaryDark"
-			app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-
-	</androidx.coordinatorlayout.widget.CoordinatorLayout>
-
-	<com.google.android.material.navigation.NavigationView
-		android:id="@+id/nav_view"
-		android:layout_width="wrap_content"
-		android:layout_height="match_parent"
-		android:layout_gravity="start"
-		android:background="@color/colorPrimaryDark"
-		android:divider="@android:color/transparent"
-		android:dividerHeight="14dp"
-		android:fitsSystemWindows="false"
-		app:headerLayout="@layout/nav_header"
-		app:itemBackground="@drawable/nav_clicked"
-		app:itemIconTint="@drawable/nav_clicked_text"
-		app:itemTextColor="@drawable/nav_clicked_text"
-		app:itemTextAppearance="@style/regular_text_white_16sp_85op"
-		app:menu="@menu/navigation_items"/>
-
-
-
-</androidx.drawerlayout.widget.DrawerLayout>
+	android:layout_width="match_parent"
+	xmlns:android="http://schemas.android.com/apk/res/android"/>
