+++ app/src/main/res/layout/fragment_places_page_item.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <!--
-  ~ Created by Andrii Kovalchuk on 01.12.19 22:42
+  ~ Created by Andrii Kovalchuk on 02.12.19 20:57
   ~ Copyright (c) 2019. All rights reserved.
-  ~ Last modified 01.12.19 22:09
+  ~ Last modified 02.12.19 20:08
   ~
   ~ This Source Code Form is subject to the terms of the Mozilla Public
   ~ License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -15,7 +15,7 @@
 	xmlns:tools="http://schemas.android.com/tools"
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
-	tools:context=".ui.main.view.MainActivity">
+	tools:context=".ui.MainActivity">
 
 	<androidx.recyclerview.widget.RecyclerView
 		android:id="@+id/content_main_rv_feed"
