--- app/src/main/res/values-v28/styles.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-
-	<style name="AppTheme" parent="Theme.AppCompat.DayNight">
-
-		<item name="android:windowLayoutInDisplayCutoutMode">never</item>
-
-	</style>
-
-
-	<style name="AppTheme.NoActionBar">
-		<item name="windowActionBar">false</item>
-		<item name="windowNoTitle">true</item>
-		<item name="android:windowDrawsSystemBarBackgrounds">true</item>
-		<item name="android:windowTranslucentStatus">true</item>
-	</style>
-
-</resources>
